+++ CHANGELOG.md
@@ -37,6 +37,7 @@
     - sets the stage for next round of `CanvasController` updates by adding `CANVAS_NAME` enum to be used when creating canvas elements
 - Add airport guide files and start ksea as an example [#29](https://github.com/openscope/openscope/issues/29)
 - Ensure tutorial selects departure aircraft [#354](https://github.com/openscope/openscope/issues/354)
+- Add documentation and tutorial section for scope commands [#718](https://github.com/openscope/openscope/issues/718)